requirements.txt
streamlit
google-generativeai
google-api-python-client
gspread
oauth2client